requirements.txt
# general modules
django >3.2,<3.3
python-decouple
pillow
djangorestframework


# third party modules
markdown
django-filter
drf-yasg[validation]


# deployment modules

# email send third party modules
django-mail-templated


#reformating tests
black
flake8